add script for automation and new redirects
