+++ .gitignore
@@ -81,5 +81,5 @@ notes.md
 reference.json
 
 ref/
-src/pages/clean.js
+newRedirects.json
 .wordpress-cache
